removed a function in app.js to clean up and instead added it to a new file called movieLoader
