+++ src/js/cmsAdapter.js
@@ -1,34 +1,30 @@
 const API_MOVIES = 'https://plankton-app-xhkom.ondigitalocean.app/api';
 
 const cmsAdapter = {
-  // Original Reviews Methods - copy from Kirill 
 
-  loadAllReviews: async () => {
-    const url = API_MOVIES + '/reviews';
-    const resp = await fetch(url);
-    const payload = await resp.json();
-    return payload.data;
+  loadMovies: async () => {
+    const url = API_MOVIES + '/movies';
+    try {
+      const resp = await fetch(url, { cache: 'no-store' });
+
+      if (!resp.ok) {
+        console.error("Error fetching movies", resp.status);
+        return [];
+      }
+
+      const fetchedMovies = await resp.json();
+
+      if (fetchedMovies && Array.isArray(fetchedMovies.data)) {
+        return fetchedMovies.data; 
+      } else {
+        console.error("Invalid data format:", fetchedMovies);
+        return [];
+      }
+    } catch (error) {
+      console.error("Error during fetch:", error);
+      return [];
+    }
   },
-
-  postReview: async (data) => {
-    const res = await fetch(API_MOVIES + '/reviews', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(data),
-    });
-    return res.json();
-  },
-
-  loadReviewById: async (id) => {
-    const url = API_MOVIES + '/reviews?filters[movie]=' + id;
-    const resp = await fetch(url);
-    const payload = await resp.json();
-    return payload.data;
-  },
-
-  // Jenny's Screenings
   
 loadAllScreenings: async () => {
     const url = API_MOVIES + '/screenings';
@@ -57,7 +53,6 @@ loadAllScreenings: async () => {
     }
   },
 
-  // Get screenings by movie ID (GET /screenings?filters[movie]=id)
   loadScreeningsByMovieId: async (id) => {
     const url = API_MOVIES + `/screenings?filters[movie]=${id}`;
     const resp = await fetch(url);
